Introduce a discovery service that wraps all actions and registration
The goal of the discovery service is to inspect every type it can find
to find explicitly owned objects. In the case where objects are
dynamically registered through actions, we inspect all the actions to
figure out how the action affects the "element tree". Basically, the
service will preemptively build a list of discovered elements. Those
elements won't yet be registered but instead we will record how those
object will "appear" to Gradle (on known, on realized, on finalized).
When we need the object to "appear" we simply need to realize the object
path. This process relies on the idea that the "element tree" is
acyclic, and only "points down". Any objects that points up would
normally be disallowed. We can point sideways.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>