First draft on the revamp of the discovery mechanic
This draft has multiple left over work which will be addressed later but
the core of the discovery mechanic will stay the same (it took about 7
iterations to come to this implementation). The general idea is to
use opaque rules which provide the candidate elements that "may" or "may
not" exists. Some candidate elements have a high degree of confident
(static), while others may depend on the parent element (select), while
others are very abstract (dynamic). The discovery "rules" are extracted
from the types and the actions which guide the discovery. The discovery
process itself if quite heavy as, in theory, any number of elements can
be nested requiring a full rebuild of the entire candidate tree at every
iteration. We could build indexes to cut down on the work but that is
mostly an implementation details which we can address later. Because
each "rules" are opaque, we can provide custom implementation type such
as the variant discovery rule. At the moment, each type and action must
be annotated. Later we could inspect the code itself.

Signed-off-by: Daniel Lacasse <daniel@nokee.dev>